Move global objects to `src/globals/`
