+++ src/globals/global-event-manager.ts
