+++ src/globals/global-manifest.ts
