+++ src/globals/global-scene.ts
