+++ src/globals/global-species-data-registry.ts
