+++ src/timed-event-manager.ts
@@ -1,6 +1,8 @@
+import { globalScene } from "#app/global-scene";
 import { speciesDataRegistry } from "#app/global-species-data-registry";
 import { SHINY_CATCH_RATE_MULTIPLIER } from "#balance/rates";
 import { CLASSIC_CANDY_FRIENDSHIP_MULTIPLIER } from "#balance/starters";
+import { timedEvents } from "#balance/timed-events";
 import type { PokemonSpeciesFilter } from "#data/pokemon-species";
 import type { MysteryEncounterTier } from "#enums/mystery-encounter-tier";
 import type { MysteryEncounterType } from "#enums/mystery-encounter-type";
@@ -10,8 +12,6 @@ import type { ModifierTypeKeys } from "#modifiers/modifier-type";
 import type { EventEncounter, EventMysteryEncounterTier, EventWeatherPools, TimedEvent } from "#types/game-events";
 import { randSeedShuffle } from "#utils/common";
 import i18next from "i18next";
-import { timedEvents } from "./data/balance/timed-events";
-import { globalScene } from "./global-scene";
 
 export class TimedEventManager {
   /**
